requirements.txt
langchain
requests
python-dotenv
langchain-groq